#19: Add "/logout" component, reorganise files
