+++ src/App.tsx
@@ -1,15 +1,14 @@
 import './App.css'
 import OpenPage from "./pages/OpenPage.tsx";
 import {Route, Routes} from "react-router-dom";
-import Layout from "./components/Layout.tsx";
+import Layout from "./components/layout/Layout.tsx";
 import Registration from "./pages/Registration.tsx";
 import Login from "./pages/Login.tsx";
 import Notes from "./pages/Notes.tsx";
 import Unauthorized from "./pages/Unauthorized.tsx";
 import MissingRoute from "./pages/MissingRoute.tsx";
-// import RequireAuth from "./components/RequireAuth.tsx";
 import Admin from "./pages/Admin.tsx";
-import AuthorizeView from "./components/AuthorizeView.tsx";
+import AuthorizeView from "./components/auth/AuthorizeView.tsx";
 
 function App() {
