+++ src/components/common/LoadingBox.tsx
