+++ src/pages/InternalError.tsx
@@ -5,7 +5,7 @@ import {
     Typography
 } from "@mui/material";
 import {Link} from "react-router-dom";
-import MainContainer from "../components/MainContainer.tsx";
+import MainContainer from "../components/common/MainContainer.tsx";
 
 const ErrorBox = styled(Box)(({theme}) => ({
     display: "flex",
