+++ src/utils/Constants.tsx
@@ -1,5 +1,7 @@
 const API_URL = {
+    MAINPAGE_URL: "/",
     LOGIN_URL: "/login",
+    LOGOUT_URL: "/api/account/logout",
     USE_COOKIES: "?useCookies=true",
     USE_SESSION_COOKIES: "?useSessionCookies=true",
     CHECK_AUTH: "/api/account/checkAuth",
